+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Fixes [#59]: Removedb name prefix when reading whitelist
 - Fixes [#62]: Provide better error messages on whitelist editing errors.
 - Fixes [#63]: Allow explicit whitelisting of database names that would be considered derivative during db cleanup
+- Fixes [#64]: Remove VISUAL and EDITOR from editor choices for DB whitelist editing
 
 ## 2.0.0 2019-02-20